Fix optionalArg check being too lenient
The original check for the optional argument (at this point the optional
argument can only be for --author or equivalently, -u) simply checked
the truthiness of optionalArg. This does well to catch undefineds, but
it also meant that any arbitrary string that was the second argument to
cli.js was considered a valid optionalArg for showing the author line.

Furthermore, this approach is not maintainable/flexible for the future,
since if any new optional arguments are added to kote, then this would
always show the author even if the new optional argument calls for
something else.